refactor(filter): Centralize all pattern normalization
